pyproject.toml
[tool.poetry]
name = "hdr-reconstruction"
version = "0.1.0"
description = ""
authors = ["kutakieu <kutakieu22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
opencv-python = "^4.8.0.76"
hydra-core = "^1.3.2"
lightning = "^2.0.8"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
wandb = "^0.16.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
